HW-1: relocate homework 1 classes
